Explicitly use long with boost::posix_time
lib/libuhd.so.3.12: undefined reference to `gps_ctrl_impl::GPSDO_COMMAND_DELAY_MS'
lib/libuhd.so.3.12: undefined reference to `uhd::usrprio_rpc::usrprio_rpc_client::DEFAULT_TIMEOUT_IN_MS'
lib/libuhd.so.3.12: undefined reference to `uhd::niusrprio::niusrprio_session::FPGA_READY_TIMEOUT_IN_MS'

host/lib/usrp/cores/rx_vita_core_3000.cpp:67:35: error: no matching conversion for functional-style cast from 'double' to 'boost::posix_time::milliseconds' (aka 'subsecond_duration<boost::posix_time::time_duration, 1000>')
        boost::this_thread::sleep(boost::posix_time::milliseconds(1.0));
                                  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
host/lib/usrp/x300/x300_impl.cpp:1533:66: error: no matching conversion for functional-style cast from 'double' to 'boost::posix_time::milliseconds' (aka 'subsecond_duration<boost::posix_time::time_duration, 1000>')
    boost::system_time timeout_time = boost::get_system_time() + boost::posix_time::milliseconds(timeout * 1000.0);
                                                                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
host/examples/rx_samples_to_file.cpp:177:64: error: no matching conversion for functional-style cast from 'double' to 'boost::posix_time::seconds'
                (boost::get_system_time() > (first_lock_time + boost::posix_time::seconds(setup_time))))
                                                               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
host/examples/rx_samples_to_file.cpp:191:53: error: no matching conversion for functional-style cast from 'double' to 'boost::posix_time::seconds'
            if (boost::get_system_time() > (start + boost::posix_time::seconds(setup_time))){
                                                    ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
host/examples/tx_samples_from_file.cpp:200:39: error: no matching conversion for functional-style cast from 'double' to 'boost::posix_time::milliseconds' (aka 'subsecond_duration<boost::posix_time::time_duration, 1000>')
            boost::this_thread::sleep(boost::posix_time::milliseconds(delay));
                                      ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/usr/local/include/boost/date_time/time_duration.hpp:270:30: note: candidate constructor (the implicit copy constructor) not viable: no known conversion from 'double' to 'const boost::date_time::subsecond_duration<boost::posix_time::time_duration, 1000>' for 1st argument
  class BOOST_SYMBOL_VISIBLE subsecond_duration : public base_duration
                             ^
/usr/local/include/boost/date_time/time_duration.hpp:270:30: note: candidate constructor (the implicit move constructor) not viable: no known conversion from 'double' to 'boost::date_time::subsecond_duration<boost::posix_time::time_duration, 1000>' for 1st argument
/usr/local/include/boost/date_time/time_duration.hpp:286:59: note: candidate template ignored: disabled by 'enable_if' [with T = double]
                                typename boost::enable_if<boost::is_integral<T>, void>::type* = 0) :
                                                          ^

+++ host/lib/transport/nirio/niusrprio_session.cpp
@@ -207,7 +207,7 @@ nirio_status niusrprio_session::_ensure_fpga_ready()
         //there is a small chance that the server is still finishing up cleaning up
         //the DMA FIFOs. We currently don't have any feedback from the driver regarding
         //this state so just wait.
-        boost::this_thread::sleep(boost::posix_time::milliseconds(FPGA_READY_TIMEOUT_IN_MS));
+        boost::this_thread::sleep(boost::posix_time::milliseconds(long(FPGA_READY_TIMEOUT_IN_MS)));
 
         //Disable all FIFOs in the FPGA
         for (size_t i = 0; i < _lvbitx->get_input_fifo_count(); i++) {
